+++ db/migrate/20210406215234_add_currency_to_pay_charges.rb
@@ -1,0 +1,5 @@
+class AddCurrencyToPayCharges < ActiveRecord::Migration[4.2]
+  def change
+    add_column :pay_charges, :currency, :string
+  end
+end