+++ lib/pay/paddle/webhooks/subscription_payment_succeeded.rb
@@ -30,8 +30,10 @@ module Pay
           params = {
             amount: Integer(event["sale_gross"].to_f * 100),
             card_type: event["payment_method"],
+            created_at: Time.zone.parse(event["event_time"]),
+            currency: event["currency"],
             paddle_receipt_url: event["receipt_url"],
-            created_at: Time.zone.parse(event["event_time"])
+            subscription: Pay::Subscription.find_by(processor: :paddle, processor_id: event["subscription_id"])
           }
 
           payment_information = Pay::Paddle::Billable.new(user).payment_information(event["subscription_id"])