+++ test/pay/paddle/webhooks/subscription_payment_succeeded_test.rb
@@ -19,6 +19,18 @@ class Pay::Paddle::Webhooks::SubscriptionPaymentSucceededTest < ActiveSupport::T
     assert_equal @data["receipt_url"], charge.paddle_receipt_url
   end
 
+  test "paddle charge is associated with subscription" do
+    @user = User.create!(email: "gob@bluth.com", processor: :paddle, processor_id: @data["user_id"])
+    subscription = @user.subscriptions.create!(name: "default", processor: :paddle, processor_id: @data["subscription_id"], processor_plan: "default", status: "active")
+
+    assert_difference "Pay.charge_model.count" do
+      PaddlePay::Subscription::User.stubs(:list).returns(:nil)
+      Pay::Paddle::Webhooks::SubscriptionPaymentSucceeded.new.call(@data)
+    end
+
+    assert_equal subscription, Pay.charge_model.last.subscription
+  end
+
   test "a charge is created and card details are set" do
     @user = User.create!(email: "gob@bluth.com", processor: :paddle, processor_id: @data["user_id"])
 
@@ -47,6 +59,7 @@ class Pay::Paddle::Webhooks::SubscriptionPaymentSucceededTest < ActiveSupport::T
     assert_equal "0020", charge.card_last4
     assert_equal "06", charge.card_exp_month
     assert_equal "2022", charge.card_exp_year
+    assert_equal "USD", charge.currency
   end
 
   test "a charge is created and paypal details are set" do