+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Unreleased
 
+* Add `Pay::Merchant` and Stripe Connect functionality - @excid3
+* Save `currency` on Pay::Charge records - @excid3
+
 ### 2.6.11
 
 * Add `subscription` method to payment processor classes for direct access to the processor subscription object.